+++ src/components/spartan/STab/STab.stories.ts
@@ -1,35 +1,71 @@
 import STab from './STab.vue';
 import STabItem from './STabItem.vue';
-import { buildSourceBinding, createDefault, createVariation } from '@/helpers';
+import { buildSourceBinding, createDefault, createVariation, createHistory } from '@/helpers';
 import { BuildingOfficeIcon, CreditCardIcon, UserIcon, UsersIcon } from '@heroicons/vue/20/solid';
 import { ref } from 'vue';
 
 export default {
     component: STab,
     title: 'navigation/Tab',
-    parameters: {
-        docs: {
-            description: {
-                component: 'The link component is used to navigate between pages.',
+    ...createHistory({
+        description: 'The tab component is used to navigate between pages.',
+        slots: [
+            {
+                name: 'default',
+                description: 'The content of the tab container.',
+            }
+        ],
+        props: [
+            {
+                name: 'modelValue',
+                description: 'The value of the tab container (variant).',
+                type: 'underline | pills | vetches',
+                default: 'undefined',
+                control: null
             },
-        },
-    },
-    argTypes: {
-        // Slots
-        default: {
-            control: { type: null },
-            table: { type: { summary: null }, category: 'Slots' },
-            description: 'The content of the tab container.',
-        },
-
-        // Props
-        variant: {
-            control: 'select',
-            options: ['underline', 'pills', 'vetches'],
-            table: { type: { summary: 'underline | pills | vetches' }, category: 'Props' },
-            description: 'The variant of the tab container.',
-        },
-    },
+            {
+                name: 'full',
+                description: 'The full width mode of the tab container.',
+                type: 'boolean',
+                default: 'false',
+            },
+            {
+                name: 'variant',
+                description: 'The variant of the tab container.',
+                options: ['underline', 'pills', 'vetches'],
+                type: 'underline | pills | vetches',
+                default: 'underline',
+            },
+            {
+                name: 'path',
+                description: 'The path of the tab item.',
+                type: 'innerText',
+                default: 'undefined',
+                subcategory: 'STabItem',
+            },
+            {
+                name: 'as',
+                description: 'The element type of the tab item.',
+                type: 'string',
+                default: 'button',
+                subcategory: 'STabItem',
+            },
+            {
+                name: 'icon',
+                description: 'The icon of the tab item.',
+                type: 'FunctionalComponent',
+                default: 'undefined',
+                subcategory: 'STabItem',
+            }
+        ],
+        events: [
+            {
+                name: 'update:modelValue',
+                description: 'The event emitted when the tab container value changes.',
+                type: 'underline | pills | vetches',
+            }
+        ]
+    }),
 };
 
 const tabs = [
@@ -41,12 +77,15 @@ const tabs = [
 
 const sourceBinding = buildSourceBinding({
     prop: { variant: 'underline' },
+    check: ['full'],
 });
 
 export const Default = createDefault({
     components: { STab, STabItem },
+    containerClass: 'w-[400px]',
     args: {
         variant: 'underline',
+        full: false,
     },
     setup: () => {
         const value = ref('Tab 1');
@@ -77,6 +116,21 @@ export const Base = createVariation({
     <STabItem>Tab 1</STabItem>
     <STabItem>Tab 2</STabItem>
     <STabItem>Tab 3</STabItem>  
+</STab>`,
+});
+
+export const Full = createVariation({
+    components: { STab, STabItem },
+    containerClass: 'bg-white -m-4 p-8',
+    setup: () => {
+        const value = ref('Tab 1');
+
+        return { value };
+    },
+    template: `<STab v-model="value" full>
+    <STabItem>Tab 1</STabItem>
+    <STabItem>Tab 2</STabItem>
+    <STabItem>Tab 3</STabItem>
 </STab>`,
 });