+++ src/components/spartan/STab/variants/underline/UnderlineTab.vue
@@ -1,8 +1,10 @@
-<script setup lang="ts"></script>
+<script setup lang="ts">
+defineProps<{ full?: boolean }>();
+</script>
 
 <template>
     <nav class="border-b border-gray-200" aria-label="Tabs">
-        <ul class="-mb-px flex space-x-8">
+        <ul :class="['-mb-px flex', { 'space-x-8': !full }]">
             <slot />
         </ul>
     </nav>