+++ src/components/spartan/STab/variants/vetch/VetchTab.vue
@@ -1,8 +1,10 @@
-<script setup lang="ts"></script>
+<script setup lang="ts">
+defineProps<{ full?: boolean }>();
+</script>
 
 <template>
     <nav aria-label="Tabs">
-        <ul class="flex space-x-2 bg-gray-100 w-fit p-2 border border-gray-300 rounded-xl">
+        <ul :class="['flex space-x-2 rounded-xl border border-gray-300 bg-gray-100 p-2', { 'w-fit': !full }]">
             <slot />
         </ul>
     </nav>