+++ CHANGELOG.md
@@ -6,7 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 2.0.0-beta.67 - 2024-02-12
+### Added
+- `SSectionDescription` now supports class merging.
+
+## 2.0.0-beta.66 - 2024-02-12
 ### Fixed
+- `SSectionTitle` now supports class merging.
+- `STab` underline full style variant.
 - `SDataTable` props documentation.
 
 ## 2.0.0-beta.65 - 2024-02-08